rename and clean custom formatting
